+++ Source/FlowEventManager/Public/FlowEventManagerComponent.h
@@ -111,7 +111,7 @@ private:
 
 	int32 ResolveNextNodeIndex(const FFlowEventNode& Node, int32 CurrentNodeIndex) const;
 	void ResolveTargets(const FFlowEventNode& Node, TArray<AActor*>& OutTargets) const;
-	bool ExecuteEventOnTarget(AActor* Target, const FFlowEventNode& Node, float OutputValue) const;
+	bool ExecuteEventOnTarget(AActor* Target, const FFlowEventNode& Node, float OutputValue, float ElapsedTime) const;
 	float EvaluateTimelineValue(const FFlowEventNode& Node, float ElapsedTime) const;
 
 	UPROPERTY()